live version at https://tw-fact-xzh-2004080613.herokuapp.com/
displays a load of debug data
count for the no. of unique IPs that searched
1. the tweet user
2. the topic
3. the tweet user + topic combination
We can show these as cool stats on the bottom such as:
Most searched User: Donald Trump, searched 1000 times
Most searched Topic: Coronavirus, searched 500 times
Most searched combination: Squidward Tentacles + Krabby Patties, searched 50 times

After we get the twitter API to display the relevant Tweets, I concatenate the tweet author, the first 50 strings of the tweet, and the timestamp to create the key for which to enter into the db. When a user selects a tweet to display news to check, this number will be incremented by one. We can use this count for each tweet to move their position up. Machine learning lol

See the live version of this at
https://tw-fact-xzh-2004080613.herokuapp.com/